+++ dropwizard-guicey/src/test/groovy/ru/vyarus/dropwizard/guice/test/jupiter/debug/HookObjectsLogTest.java
@@ -189,6 +189,6 @@ public class HookObjectsLogTest extends AbstractPlatformTest {
 
     @Override
     protected String clean(String out) {
-        return out.replaceAll("\\d+\\.\\d+ ms", "111 ms");
+        return out.replaceAll("\\d+(\\.\\d+)? ms", "111 ms");
     }
 }