+++ dropwizard-guicey/src/test/groovy/ru/vyarus/dropwizard/guice/test/jupiter/debug/NestedConfigOverrideLogTest.java
@@ -64,6 +64,6 @@ public class NestedConfigOverrideLogTest extends AbstractPlatformTest {
 
     @Override
     protected String clean(String out) {
-        return out.replaceAll("\\d+\\.\\d+ ms", "111 ms");
+        return out.replaceAll("\\d+(\\.\\d+)? ms", "111 ms");
     }
 }